pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[project]
name = "dagpiler"
version = '0.1.2'
readme = "README.md"
license ={ file = "LICENSE" }
description = 'Compile a DAG representation of a data processing pipeline from TOML files.'
authors = [{name = "Mitchell Tillman", email ="mtillman14@gmail.com"}]
dependencies = [
    "toml",
    "networkx",
    "requests",    
    # Documentation dependencies
    "mkdocs-material",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-awesome-pages-plugin",
    "mkdocstrings",
    "pyyaml",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",    
    "Operating System :: OS Independent",
]
[project.urls]
Homepage = "https://github.com/ResearchOS/dagpiler"
Documentation = "https://researchos.github.io/dagpiler/"
Repository = "https://github.com/ResearchOS/dagpiler"
Issues = "https://github.com/ResearchOS/dagpiler/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[dependency-groups]
tests = [
    "pytest",
]
plot = [
    "matplotlib",
]

[project.scripts]
dagpiler = "dagpiler.__main__:main"

[tool.hatch.build.targets.sdist]
include = "src"

[tool.hatch.build.targets.wheel]
packages = [
    "src/dagpiler"
]